requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
Flask-Migrate
psycopg2-binary
python-dotenv